+++ libs/ethers/src/lib/dao/components/dao-signer-form/dao-signer-form.component.ts
@@ -1,29 +1,28 @@
 import { ControlContainer } from '@angular/forms';
 import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-import { OrganizationOperation, OrganizationService, OrganizationQuery } from '../../+state';
 import { MatSlideToggleChange } from '@angular/material';
-import { PermissionsQuery } from '../../permissions/+state';
 import { Router } from '@angular/router';
 import { WalletService } from 'libs/ethers/src/lib/wallet/+state';
 import { ActionTx, TxFeedback } from '@blockframes/ethers/types';
-import { CreateTx } from '@blockframes/ethers';
-import { OrganizationMember } from '../../member/+state/member.model';
+import { CreateTx } from '@blockframes/ethers/create-tx';
+import { OrganizationMember } from '@blockframes/organization/member/+state/member.model';
+import { PermissionsQuery } from '@blockframes/organization/permissions/+state';
+import { DaoService, DaoQuery, DaoOperation } from '../../+state';
 
 @Component({
-  selector: 'organization-signer-form',
-  templateUrl: './organization-signer-form.component.html',
-  styleUrls: ['./organization-signer-form.component.scss'],
+  selector: 'dao-signer-form',
+  templateUrl: './dao-signer-form.component.html',
+  styleUrls: ['./dao-signer-form.component.scss'],
   changeDetection: ChangeDetectionStrategy.Default
 })
-export class OrganizationSignerFormComponent {
+export class DaoSignerFormComponent {
 
   @Input() member: OrganizationMember;
 
   constructor(
     public controlContainer: ControlContainer,
     private permissionQuery: PermissionsQuery,
-    private service: OrganizationService,
-    private query: OrganizationQuery,
+    private service: DaoService,
     private router: Router,
     private walletService: WalletService,
   ) { }
@@ -36,13 +35,13 @@ export class OrganizationSignerFormComponent {
     return this.permissionQuery.isUserAdmin(this.member.uid);
   }
 
-  isSelected(operation: OrganizationOperation) {
+  isSelected(operation: DaoOperation) {
     return operation.members.some(operationMember => operationMember.uid === this.member.uid) || this.isAdmin;
   }
 
   public async toggleSelection(toggle: MatSlideToggleChange, id: string) {
-    const operations: OrganizationOperation[] = this.control.value.filter(operation => operation.id !== id);
-    const currentOperation: OrganizationOperation = this.control.value.find(operation => operation.id === id);
+    const operations: DaoOperation[] = this.control.value.filter(operation => operation.id !== id);
+    const currentOperation: DaoOperation = this.control.value.find(operation => operation.id === id);
     const members: OrganizationMember[] = currentOperation.members.filter(operationMember => operationMember.uid !== this.member.uid);
     if (toggle.checked) {
       members.push(this.member);
@@ -54,7 +53,6 @@ export class OrganizationSignerFormComponent {
     const orgEthAddress = await this.service.getOrganizationEthAddress();
     const memberName = this.member.name;
     const operationName = currentOperation.name;
-    const orgId = this.query.getActiveId();
 
     let tx: ActionTx;
     let feedback: TxFeedback;
@@ -64,7 +62,7 @@ export class OrganizationSignerFormComponent {
         confirmation: `You are about to whitelist ${memberName} for ${operationName}`,
         success: `${memberName} has been successfully whitelisted !`,
         redirectName: 'Back to Administration',
-        redirectRoute: `/layout/o/organization/${orgId}/administration`,
+        redirectRoute: `/layout/o/organization/`,
       }
     } else {
       tx = CreateTx.removeMember(orgEthAddress, id, memberEthAddress);
@@ -72,7 +70,7 @@ export class OrganizationSignerFormComponent {
         confirmation: `You are about to blacklist ${memberName} for ${operationName}`,
         success: `${memberName} has been successfully blacklisted !`,
         redirectName: 'Back to Administration',
-        redirectRoute: `/layout/o/organization/${orgId}/administration`,
+        redirectRoute: `/layout/o/organization/`,
       }
     }
