+++ libs/ethers/src/lib/dao/guards/dao.guard.ts
@@ -1,0 +1,21 @@
+import { Injectable } from "@angular/core";
+import { CollectionGuardConfig } from "akita-ng-fire";
+import { DaoService } from "../+state";
+
+@Injectable({ providedIn: 'root' })
+@CollectionGuardConfig({ awaitSync: true })
+export class DaoGuard {
+  constructor(
+    private service: DaoService
+  ) { }
+
+  canActivate() {
+    this.service.retrieveDataAndAddListeners();
+    return true;
+  }
+
+  canDeactivate() {
+    this.service.removeAllListeners();
+    return true;
+  }
+}
