+++ libs/ethers/src/lib/dao/pages/dao-activity-view/dao-activity-view.component.scss
