+++ libs/ethers/src/lib/dao/pages/dao-activity-view/dao-activity-view.component.ts
@@ -1,28 +1,28 @@
 import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { OrganizationAction, OrganizationQuery } from '../../+state';
+import { DaoAction, DaoQuery } from '../../+state';
 
 @Component({
-  selector: 'org-activity-view',
-  templateUrl: './organization-activity-view.component.html',
-  styleUrls: ['./organization-activity-view.component.scss'],
+  selector: 'dao-activity-view',
+  templateUrl: './dao-activity-view.component.html',
+  styleUrls: ['./dao-activity-view.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class OrganizationActivityViewComponent implements OnInit {
+export class DaoActivityViewComponent implements OnInit {
 
   /** Observable which contains pending actions */
-  public pendingActions$ = new Observable<OrganizationAction[]>();
+  public pendingActions$ = new Observable<DaoAction[]>();
 
   /** Observable which contains approved actions */
-  public approvedActions$ = new Observable<OrganizationAction[]>();
+  public approvedActions$ = new Observable<DaoAction[]>();
 
-  public selectedAction: OrganizationAction;
+  public selectedAction: DaoAction;
 
   // Flag to indicate whether sidenav is open or not
   public opened = false;
 
   constructor(
-    private query: OrganizationQuery
+    private query: DaoQuery
   ){}
 
   ngOnInit() {
@@ -30,9 +30,8 @@ export class OrganizationActivityViewComponent implements OnInit {
     this.approvedActions$ = this.query.approvedActions$;
   }
 
-
   // also the sidenav is going to need a refactoring
-  public openSidenav(action: OrganizationAction) {
+  public openSidenav(action: DaoAction) {
     this.opened = true;
     this.selectedAction = action;
   }
