+++ libs/ethers/src/lib/dao/pages/dao-admin-view/dao-admin-view.component.html
@@ -21,15 +21,15 @@
         <!-- Operations -->
         <section>
           <h3>Operations</h3>
-          <org-display-operations [operations]="operations" (editing)="openSidenav('operation', $event)" >
-          </org-display-operations>
+          <dao-display-operations [operations]="operations" (editing)="openSidenav('operation', $event)" >
+          </dao-display-operations>
         </section>
 
         <!--Operation Permissions-->
         <section>
           <h3>Users Operations Permissions</h3>
-          <org-signer-repertory [members]="members" [operations]="operations" (selected)="openSidenav('member', $event)" >
-          </org-signer-repertory>
+          <dao-signer-repertory [members]="members" [operations]="operations" (selected)="openSidenav('member', $event)" >
+          </dao-signer-repertory>
         </section>
 
       </ng-container>
@@ -39,7 +39,7 @@
             <ng-container *ngIf="(operationFormGroup$ | async) as operationFormGroup" >
 
               <!-- Operation -->
-              <organization-form-operation [formGroup]="operationFormGroup" [members]="members"></organization-form-operation>
+              <dao-form-operation [formGroup]="operationFormGroup" [members]="members"></dao-form-operation>
 
             </ng-container>
           </ng-container>
@@ -47,7 +47,7 @@
             <ng-container *ngIf="(memberFormGroup$ | async) as memberFormGroup" >
 
               <!-- Members operation permission -->
-              <organization-signer-form [formGroup]="operationFormList" [member]="memberFormGroup.value"></organization-signer-form>
+              <dao-signer-form [formGroup]="operationFormList" [member]="memberFormGroup.value"></dao-signer-form>
 
             </ng-container>
           </ng-container>
