+++ libs/ethers/src/lib/dao/tslint.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../tslint.json",
+  "rules": {
+    "directive-selector": [true, "attribute", "dao", "camelCase"],
+    "component-selector": [true, "element", "dao", "kebab-case"]
+  }
+}
